fix and refactor files more
