+++ Behavioral/ChainOfResponsibility/index.html
@@ -1,33 +1,1 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width,initial-scale=1" />
-    <title>Chain of Responsibility Pattern</title>
-    <link rel="stylesheet" href="style.css" />
-    <script defer="defer" src="dist/index.js"></script>
-  </head>
-  <body>
-    <div class="window">
-      <div class="form-container">
-        <div class="form-title"><h4>Login</h4></div>
-        <form id="validationForm" class="validationForm">
-          <div class="form-input">
-            <label for="username">Username:</label>
-            <input id="username" placeholder="Username" />
-          </div>
-          <div class="form-input">
-            <label for="email">Email:</label>
-            <input type="email" id="email" placeholder="Email" />
-          </div>
-          <div class="form-input">
-            <label for="password">Password:</label>
-            <input type="password" id="password" placeholder="Password" />
-          </div>
-          <button type="submit">Submit</button>
-        </form>
-      </div>
-    </div>
-  </body>
-</html>
+<!doctype html><html lang="en"><head><meta charset="UTF-8"/><meta http-equiv="X-UA-Compatible" content="IE=edge"/><meta name="viewport" content="width=device-width,initial-scale=1"/><title>Chain of Responsibility Pattern</title><link rel="stylesheet" href="style.css"/><script defer="defer" src="dist/index.js"></script><script defer="defer" src="dist/index.js"></script></head><body><div class="window"><div class="form-container"><div class="form-title"><h4>Login</h4></div><form id="validationForm" class="validationForm"><div class="form-input"><label for="username">Username:</label> <input id="username" placeholder="Username"/></div><div class="form-input"><label for="email">Email:</label> <input type="email" id="email" placeholder="Email"/></div><div class="form-input"><label for="password">Password:</label> <input type="password" id="password" placeholder="Password"/></div><button type="submit">Submit</button></form></div></div></body></html>
