+++ Behavioral/ChainOfResponsibility/src/interfaces/BaseHandler.ts
