+++ Behavioral/ChainOfResponsibility/src/interfaces/ValidationHandler.ts
