+++ Behavioral/ChainOfResponsibility/src/interfaces/ValidationInput.ts
