+++ Behavioral/ChainOfResponsibility/src/models/EmailFormatHandler.ts
@@ -1,5 +1,5 @@
-import BaseHandler from "./BaseHandler";
-import ValidationInput from "./ValidationInput";
+import BaseHandler from "../interfaces/BaseHandler";
+import ValidationInput from "../interfaces/ValidationInput";
 
 class EmailFormatHandler extends BaseHandler {
   handleValidation(inputValue: ValidationInput): boolean {
