+++ Behavioral/Command/src/interfaces/Command.ts
@@ -1,0 +1,4 @@
+export interface Command {
+  execute(): any;
+  undo(): any;
+}
