+++ Behavioral/Command/src/interfaces/Database.ts
@@ -1,0 +1,6 @@
+import { Data } from "../types/Data";
+
+export interface Database {
+  insert(data: Data): void;
+  delete(data: Data): void;
+}
