+++ Behavioral/Command/src/types/Data.ts
@@ -1,0 +1,5 @@
+export type Data = {
+  id: number;
+  username: string;
+  information: string;
+};
