+++ Behavioral/Mediator/src/interfaces/ChatMediator.ts
