+++ Behavioral/Mediator/src/interfaces/Component.ts
@@ -1,0 +1,9 @@
+import { ChatMediator } from "./ChatMediator";
+
+export abstract class Component {
+  protected mediator: ChatMediator;
+
+  constructor(mediator: ChatMediator) {
+    this.mediator = mediator;
+  }
+}
