+++ Behavioral/Mediator/src/interfaces/User.ts
