+++ Behavioral/Observer/src/interfaces/Observer.ts
