+++ Behavioral/Observer/src/interfaces/Subject.ts
