+++ Behavioral/State/src/models/Silver.ts
@@ -1,4 +1,4 @@
-import { State } from "./State";
+import { State } from "../interfaces/State";
 import { Red } from "./Red";
 import { Gold } from "./Gold";
