+++ Behavioral/Strategy/src/interfaces/Payment.ts
@@ -1,0 +1,6 @@
+import { User } from "../models/User";
+
+// Payment.ts
+export interface Payment {
+  pay(user: User): number;
+}
