+++ Behavioral/Visitor/src/interfaces/Employee.ts
