+++ Behavioral/Visitor/src/interfaces/EmployeeVisitor.ts
