--- Behavioral/Visitor/src/models/Employees.ts
@@ -1,27 +1,0 @@
-// Employees.ts
-import { Employee } from "./Employee";
-import { EmployeeVisitor } from "./EmployeeVisitor";
-
-export class Engineer implements Employee {
-  constructor(
-    public name: string,
-    public salary: number,
-    public department: string
-  ) {}
-
-  accept(visitor: EmployeeVisitor): void {
-    visitor.visit(this);
-  }
-}
-
-export class Manager implements Employee {
-  constructor(
-    public name: string,
-    public salary: number,
-    public department: string
-  ) {}
-
-  accept(visitor: EmployeeVisitor): void {
-    visitor.visit(this);
-  }
-}
